% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\alias{ez_annotate}
\alias{ez_biotype_stat}
\title{Annotate using biomart}
\usage{
ez_annotate(x, host = "www.ensembl.org", dataset = "hsapiens_gene_ensembl")

ez_biotype_stat(count_data, biotype, samples = NULL, compact = FALSE)
}
\arguments{
\item{x}{a vector of ids, matrix or a data frame. if x is a matrix or data frame,
Then rownames is taken as ids. Ids can be Ensembl ids or gene names}

\item{dataset}{Dataset you want to use. See ?biomaRt::useMart}

\item{count_data}{a data frame containing raw read count}

\item{biotype}{the column containing biotype data of each gene}

\item{samples}{a vector of sample names for which you want to do the stat}

\item{compact}{logical for making compact statistique}
}
\description{
\itemize{
\item ez_annotate(): annotate a list of gene names or ensembl ids using biomaRt
\item ez_biotype_stat(): Make a stat summary of gene biotype
}
}
