% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_count_data.R
\name{check_count_data}
\alias{check_count_data}
\title{Normalize Counts Data Using DESeq2}
\usage{
check_count_data(data, result.dir = "COUNT")
}
\arguments{
\item{data}{a list containing the following component: raw.count, count.norm
and count.rlog. Genrally an object of class normalize_counts as generated by
the function normalize_counts().}

\item{result.dir}{a directory to save the output.}
}
\value{
Return an object of class SummarizedExperiment.

Create a pdf file "quality.control.pdf" containing the results of the
 quality assessment.
}
\description{
Normalize Counts Data Using DESeq2.
}

