% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_samples_file.R
\name{create_samples_file}
\alias{create_samples_file}
\title{Create Samples Annotation File}
\usage{
create_samples_file(fastq_dir = "FASTQ", fastq1_pattern = "_R1.fastq.gz",
  fastq2_pattern = "_R2.fastq.gz")
}
\arguments{
\item{fastq_dir}{path to the directory containing the fastq files.}

\item{fastq1_pattern}{character string in the fastq1 files}

\item{fastq2_pattern}{character string in the fastq2 files (for paired-end sequencing).}
}
\value{
Create a samples.txt file in the working directory.
}
\description{
Create samples annotation file according fastq files contained in a FASTQ directory.
}

