% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Count distribution}
\usage{
distribution(x, sample_cols = NULL, mincount = 0, groupby = NULL,
  group_cols = NULL)
}
\arguments{
\item{x}{a numeric matrix containing raw or normalized counts; columns are samples and rows are genes}

\item{sample_cols}{a vector containing color for each sample.}

\item{mincount}{positive value; only genes with count > mincount are kept.}

\item{groupby}{a factor variable specifying sample groups}

\item{group_cols}{a vector specifying the color for each group}
}
\description{
Count distribution
\itemize{
\item distribution(): plots the count distribution per sample or per group
}
}

