% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq_nb_reads.R
\name{fastq_nb_reads}
\alias{fastq_nb_reads}
\title{Count the number of reads in fastq files}
\usage{
fastq_nb_reads(fastq = NULL, ext = "fastq", show_progress = TRUE)
}
\arguments{
\item{fastq}{a vector or a directory containing one or more fastq files. if NULL, fastq files in the 
current working directory are choosed}

\item{ext}{fastq file extension. Default is "fastq". If you have paired-end sequencing data, you can use ext = "R1.fastq", 
to avoid counting twice. Note that "R1_fastq" and "R2_fastq" contain the same number of reads}

\item{show_progress}{if TRUE, the progression is shown}
}
\value{
Returns a data frame containing file name (column 1) and read count (column 2)
}
\description{
Count the number of reads in fastq files.
}
\examples{
\donttest{
fastq_nb_reads(fastq, ext = "R1.fastq")
}
}
