% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_igenome.R
\name{install_igenome}
\alias{install_igenome}
\title{Download Reference Genome from iGenome}
\usage{
install_igenome(version = NULL, path = NULL, result_dir = "~/genomes")
}
\arguments{
\item{version}{the version of the genome of interest (e.g., "hg19"). Allowed values are "hg18", "hg19" and "GRCh37".
For installing other iGenome versions, specify the path}

\item{path}{the path to the genome of interest. Not required if version = "hg18", "hg19" or "GRCh37".}

\item{result_dir}{a directory to hold the genome. Default is "~/genomes".}
}
\value{
Create a result directory containing the reference igenome.
Default directory is "~/genomes" which will be created in your home directory.
}
\description{
Download and install illumina \href{http://ccb.jhu.edu/software/tophat/igenomes.shtml}{iGenomes}.
}
\examples{
\donttest{
install_igenome(version = "grch37", result_dir = "~/genomes")
}

}
