% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_counts.R
\name{normalize_counts}
\alias{normalize_counts}
\title{Normalize Counts Data Using DESeq2}
\usage{
normalize_counts(data, thread = 10, result.dir = "COUNT", save = TRUE)
}
\arguments{
\item{data}{an object of class data.frame, matrix, SummarizedExmeriment or
DESeqDataSet containing raw counts}

\item{thread}{Number of threads to be used. This depends to the available
computer ressources.}

\item{result.dir}{a directory to save the output.}

\item{save}{logical value; if TRUE, the results are saved in result.dir.}
}
\value{
Return an object of class SummarizedExperiment.

returns a list containing the following components: \itemize{ \item
 count.norm: Normalized count for sequencing depth \item count.rlog:
 rlog-transformed data: variance stabilization. To be used for visualization
 such as PCA and clustering. }
}
\description{
Normalize Counts Data Using DESeq2.
}

