% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samples_count.R
\name{plot_samples_count}
\alias{plot_samples_count}
\title{Plot Samples Count}
\usage{
plot_samples_count(data, color = "white", fill = "steelblue",
  ylab = "Total Count (x 10^7)", xlab = "Sample", name.size = 7, ...)
}
\arguments{
\item{data}{An object of class data.frame, matrix or SummarizedExperiment.}

\item{color}{outline and fill colors.}

\item{fill}{fill color for the bars}

\item{name.size}{Change the sample name size}

\item{...}{Add other arguments from ggbarplot to modify the plot appearance.}
}
\value{
Return a plot ggplot.
}
\description{
Plot the total count per sample
}

