% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{rnaseq_eda}
\alias{rnaseq_eda}
\title{Create reports}
\usage{
rnaseq_eda(se, result_file)
}
\arguments{
\item{se}{an object of class SummarizedExperiment generated by the function count_reads()}

\item{result_file}{path to the result_file prefix (e.g., path/to/prefix)}
}
\description{
Create an html file containing some exploratory data analysis
}

