% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseq_workflow.R
\name{rnaseq_workflow}
\alias{rnaseq_workflow}
\title{RNAseq Workflow}
\usage{
rnaseq_workflow(data_dir = getwd(), samples.annotation = "samples.txt",
  pairedEnd = TRUE, fastq.gz = TRUE,
  star.index = "/eqmoreaux/genomes/Homo_sapiens/Ensembl/GRCh37/Sequence/StarIndex/",
  gtf = "/eqmoreaux/genomes/Homo_sapiens/Ensembl/GRCh37/Annotation/Genes/genes.gtf",
  result.dir = getwd(), keep = c("name_sorted_bam"),
  ignore.strand = FALSE, count.mode = "Union", thread = 10)
}
\arguments{
\item{data_dir}{the directory containing the data. It should contains a subdirectory called "FASTQ" which
contains fastq files. Default value is the working directory}

\item{samples.annotation}{the path to "samples.txt" tabulation file containing sample annotations.
This file should contain at least 3 columns:
\itemize{
\item name: corresponding to sample name. It will be used to report the result
\item fastq1: the first fastq file name for paired-end sequencing
\item fastq2: the second fastq file name for paired-end sequencing.
This is optional if single-end sequencing is used
}}

\item{pairedEnd}{specify if the data are paired-end sequencing data. Default is TRUE.}

\item{fastq.gz}{specify if the FASTQ files are compressed. Default is TRUE.}

\item{star.index}{reference genome for STAR.}

\item{gtf}{gene annotation file.}

\item{result.dir}{The absolute path of result directory.}

\item{keep}{a character vector specifying the file to be kept.
Allowed values include c("name_sorted_bam", "chr_sorted_bam", "sam" ).}

\item{ignore.strand}{A logical indicating if strand should be considered when
matching. if TRUE, allows minus strand reads to count in plus strand genes,
and vice versa. Should be FALSE for stranded RNAsequencing.}

\item{count.mode}{counting mode. Read ?GenomicAlignments::summarizeOverlaps. Default
value is "Union". Reads that overlap any portion of exactly one feature are
counted. Reads that overlap multiple features are discarded.}

\item{thread}{Number of threads to be used. This depends to the available computer ressources.}
}
\value{
Can create three subdirectories: \cr
\itemize{
\item SAM: containing the output of STAR alignment program
\item BAM: containing unsorted and sorted BAM files, and BAM index file.
BAM comtent is sorted by read names (*_name_sorted.bam) or by chromosome (*_sorted.bam).
The index files are of form *_sorted.bam.bai. *_name_sorted.bam files are used for read counting
using hseq-count or R. *_sorted.bam and *_sorted.bam.bai files can be used for IGV visualization.
\item COUNT: containing read counting results. It contains the following files:\cr
1. se.RDATA: containing an object "se" which is an object of class SummarizedExperiment
}
}
\description{
RNAseq Workflow. Alignment with STAR and read counting with R/Bioconductor.
}

