% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{Read Counting}
\usage{
count_reads(bam = NULL, ext = ".bam", by = c("gene", "exon"),
  samples.annotation = NULL, result.dir = "COUNT", save = TRUE,
  gtf = "/eqmoreaux/genomes/Homo_sapiens/Ensembl/GRCh37/Annotation/Genes/genes.gtf",
  pairedEnd = TRUE, ignore.strand = FALSE, count.mode = "Union",
  thread = 10)
}
\arguments{
\item{bam}{a vector of bam files (in the same directory) or a directory
containing one or more paths to bam files. if NULL, BAM files
in the current working directory are choosed.}

\item{ext}{BAM file extension. Default is ".bam". Used only when bam is a directory}

\item{by}{One of "gene" (for counting in gene) or "exon" (for counting in
exon)}

\item{samples.annotation}{the path to "samples.txt" tabulation file containing sample annotations.
This file should contain at least 3 columns:
\itemize{
\item name: corresponding to sample name. It will be used to report the result
\item fastq1: the first fastq file name for paired-end sequencing
\item fastq2: the second fastq file name for paired-end sequencing.
This is optional if single-end sequencing is used
}}

\item{result.dir}{a directory to save the output.}

\item{save}{logical value; if TRUE, the result is saved in result.dir.}

\item{gtf}{gene annotation file.}

\item{pairedEnd}{specify if the data are paired-end sequencing data. Default is TRUE.}

\item{ignore.strand}{A logical indicating if strand should be considered when
matching. if TRUE, allows minus strand reads to count in plus strand genes,
and vice versa. Should be FALSE for stranded RNAsequencing.}

\item{count.mode}{counting mode. Read ?GenomicAlignments::summarizeOverlaps. Default
value is "Union". Reads that overlap any portion of exactly one feature are
counted. Reads that overlap multiple features are discarded.}
}
\value{
Return an object of class SummarizedExperiment
}
\description{
Count reads by gene and by exon
}
\details{
Used Bioconductor packages: \itemize{ \item \strong{GenomicFeatures}
to prepare the transcript database \item \strong{Rsamtools} to import SAM/BAM
files in R \item \strong{GenomicAlignments} for read counting \item
\strong{BiocParallel}for parallel computing }

Input files: BAM files. No need to sort the BAM files in Bioconductor version
> 2.12.
}

