% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_reads.R
\name{count_reads}
\alias{count_reads}
\title{Count reads per genes using Bioconductor package}
\usage{
count_reads(bam = NULL, ext = "name_sorted.bam", by = c("gene", "exon"),
  result_dir = "COUNT", save = TRUE, show_progress = TRUE,
  gtf = "/eqmoreaux/genomes/Homo_sapiens/Ensembl/GRCh37/Annotation/Genes/genes.gtf",
  pairedEnd = TRUE, ignore.strand = FALSE, mode = "Union", thread = 25)
}
\arguments{
\item{bam}{a vector of bam files (in the same directory)
 or a directory containing one or more paths to bam files sorted by name.
if NULL, BAM files in the current working directory are choosed.}

\item{ext}{BAM file extension. For example ext = "name_sorted.bam".
Default is "name_sorted.bam". Used only when bam is a directory}

\item{by}{One of "gene" (for counting in gene) or "exon" (for counting in exon)}

\item{result_dir}{a directory to save the output.}

\item{save}{logical value; if TRUE, the result is saved in result_dir.}

\item{show_progress}{if TRUE, the progression is shown}

\item{gtf}{reference genome for STAR and annotation gtf file, respectively.}

\item{pairedEnd}{specify if the data are paired-end sequencing data. Default is TRUE.}

\item{ignore.strand}{A logical indicating if strand should be considered when matching.
Used during read counting.}

\item{mode}{counting mode. Read ?summarizeOverlaps}

\item{thread}{Number of threads to be used. This depends to the available computer ressources.}
}
\value{
Return an object of class SummarizedExperiment
}
\description{
Count reads per genes using Bioconductor package
}
\details{
required bioconductor packages: "GenomicFeatures", "Rsamtools", "GenomicAlignments" and "BiocParallel".
}

